dwai_api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.0
graphql-core==3.2.3
h11==0.14.0
idna==3.6
motor==3.3.2
mypy==1.9.0
mypy-extensions==1.0.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
strawberry-graphql==0.220.0
tomli==2.0.1
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
uvicorn==0.27.1
watchfiles==0.21.0
